bookworld/requirements.txt
django-extensions==3.1.5